Initial implementation of FreshVar
`FreshVar` uses a closure as an object design pattern to wrap a value
with doubly nested `MVar`, two stale checks, a refresh action, and a
mutex. The two stale checks allow the value to check when it needs to be
refreshed. This was originally designed for tokens that may expire,
however it is general enough to wrap over any pure value that may become
stale.

One design decision to note is that the stale checks are pure. This may
not work for all circumstances, i.e. stale checks that require IO.
However these types of stale checks are expensive and do not work well
for quick access to a global value.