Allow reads when trying to refresh
A `tryRefresh` should only block when it actually needs to mutate the
`FreshVar`. By wrapping it in `modifyFreshVar` it forced blocking during
the whole async refresh. Instead the mutex can be quickly read from the
`FreshVar` and utilized to prevent any concurrent updates.

This requires that the mutex value is never replaced inside of `Fresh`.